Update to BIGINT and BIGSERIAL to match Long definition in Models

+++ create.sql
@@ -108,8 +108,8 @@ GRANT ALL PRIVILEGES ON SEQUENCE account_device_map_id_seq TO ingress_user;
 
 
 CREATE TABLE account_tracker_map(
-    id SERIAL PRIMARY KEY,
-    account_id INTEGER,
+    id BIGSERIAL PRIMARY KEY,
+    account_id BIGINT,
     device_id VARCHAR(100),
     created_at TIMESTAMP default current_timestamp
 );
@@ -240,8 +240,8 @@ CREATE TABLE device_sensors_batch (
 -- Assume for now we only support one tracker each account for sleep cycle tracking
 CREATE TABLE tracker_motion_master(
     id BIGSERIAL PRIMARY KEY,
-    account_id INTEGER,
-    tracker_id INTEGER,
+    account_id BIGINT,
+    tracker_id BIGINT,
     svm_no_gravity INTEGER,
     ts TIMESTAMP,
     offset_millis INTEGER,
